+++ src/utils/promiseValueHolder.ts
@@ -18,6 +18,21 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
+export class PromiseValueHolderStore<T> {
+    private creator: () => PromiseValueHolder<T>;
+    private valueHolder: PromiseValueHolder<T>|undefined;
+    
+    constructor(creator: () => PromiseValueHolder<T>) {
+        this.creator = creator;
+    }
+    
+    getValue(): Promise<T>|T {
+        if(this.valueHolder === undefined)
+            this.valueHolder = this.creator();
+        return this.valueHolder.getValue();
+    }
+}
+
 export default class PromiseValueHolder<T> {
     private promise: Promise<T>;
     private isResolved: boolean = false;